reorganize lots of things
This now has a functional API, where each sort of object is in it's own folder.
eg there is a tables/ folder that has all the types and API methods to interact with tables, another for fields, records, and bases.

Now the client stuff is just an object oriented wrapper around this that can help reduce the number of parameters
you have to hold onto, but you can get to bare bones still if you want.

I also reworked the test infra so that bringing the test base up to speed actually works.