WIP: overhaul field value converters to require less info.
Before they were typed so that they required the entire field schema, including the field id and name. That was unneeded, all the individual converters need is the fiel type, and for some fields the options as well.

Also, the implementation was way messier than it needed to be, with some converter interface and nesting.
Now it is flatter and simpler.
The type inference is also simpler and easier to copy-paste, with there only a sinlge source of truth, the actual handler functions.
All in all I'm quite happy with these adjust ments.

I am still adjusting the rest of the library to account for this change. I think I got the record converter figured out, but still need to adjust the typings in the records api

I moved some of the exception implementations into the respective packages, but still export them from the exceptions namespace. There is some awkwardness there to prevent circular imports.